moved updater to separate namespace
